add presentation package, refactor
