+++ app/src/main/java/com/example/compnote/presentation/navigation/CompnoteNavHost.kt
@@ -1,14 +1,14 @@
-package com.example.compnote.navigation
+package com.example.compnote.presentation.navigation
 
 import androidx.compose.runtime.Composable
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
-import com.example.compnote.MainViewModel
-import com.example.compnote.screens.AddScreen
-import com.example.compnote.screens.MainScreen
-import com.example.compnote.screens.NoteScreen
-import com.example.compnote.util.Constants
+import com.example.compnote.presentation.MainViewModel
+import com.example.compnote.presentation.screens.AddScreen
+import com.example.compnote.presentation.screens.MainScreen
+import com.example.compnote.presentation.screens.NoteScreen
+import com.example.compnote.presentation.util.Constants
 
 @Composable
 fun CompnoteNavHost(mViewModel: MainViewModel) {
