+++ app/src/main/java/com/example/compnote/presentation/ui/theme/Color.kt
@@ -1,4 +1,4 @@
-package com.example.compnote.ui.theme
+package com.example.compnote.presentation.ui.theme
 
 import androidx.compose.ui.graphics.Color
